mental chatbot/backend/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.8.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
build==1.2.2.post1
cachetools==5.5.0
catboost==1.2.7
category-encoders==2.6.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.11
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
durationpy==0.9
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
fasteners==0.14.1
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
future==0.18.2
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.68.0
gradio==5.19.0
gradio_client==1.7.2
graphviz==0.20.3
greenlet==3.1.1
groq==0.16.0
grpcio==1.69.0
gyp==0.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.3
importlib_metadata==8.5.0
importlib_resources==6.5.2
instagram-story==1.3.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.13.1
Keras-Applications==1.0.8
keras-vggface @ git+https://github.com/ma7555/keras-vggface.git@97a512377f9ead550a74595b657c7ec9b8161e75
keyring==23.5.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-groq==0.2.4
langchain-text-splitters==0.3.6
langgraph==0.3.1
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.3.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
loguru==0.7.3
louis==3.20.0
lxml==5.3.0
lz4==4.3.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.4.1
mmh3==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.24.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
newsapi-python==0.2.7
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.20.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
patsy==1.0.1
peewee==3.17.8
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
plotly==6.0.0
pmdarima==2.0.3
posthog==3.17.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
PuLP==2.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==5.3.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.1
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.13.5
streamlit==1.41.1
sympy==1.13.1
systemd-python==234
ta==0.11.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.49.0
triton==3.1.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.1
xdg==5
xgboost==1.7.3
xkit==0.0.0
xlrd==2.0.1
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
zstandard==0.23.0
